Move Person Form tests from People.test.tsx to PersonForm.test.tsx
